requirements.txt
Django
line-bot-sdk
Pillow
django-environ